+++ src/components/index.ts
@@ -34,6 +34,7 @@ import PostButton from "./commons/Button/PostButton/PostButton";
 import DatePickerButton from "./commons/DatePicker/DatePickerButton/DatePickerButton";
 import NoData from "./commons/NoData/NoData";
 import StateButton from "./commons/Button/StateButton/StateButton";
+import Pagination from "./commons/Pagination/Pagination";
 
 export {
   Modal,
@@ -72,4 +73,5 @@ export {
   DatePickerButton,
   NoData,
   StateButton,
+  Pagination,
 };
