+++ src/pages/Home/components/PopularPots/PopularPots.style.ts
@@ -19,20 +19,3 @@ export const buttonContainer = css`
   justify-content: center;
   align-items: center;
 `;
-
-export const pageNumberStyle = css`
-  ${theme.font.caption3}
-  color: ${theme.color.point.gray};
-`;
-
-export const buttonStyle = css`
-  padding: 1rem;
-  border-radius: 50%;
-  border: none;
-  background-color: ${theme.color.point.normal};
-  height: 4.4rem;
-`;
-
-export const iconStyle = css`
-  transform: scaleX(-1);
-`;
