+++ src/pages/Home/components/PopularPots/PopularPots.tsx
@@ -1,19 +1,13 @@
 import { Swiper, SwiperSlide } from "swiper/react";
 import { Navigation, Pagination } from "swiper/modules";
-import {
-  buttonContainer,
-  buttonStyle,
-  iconStyle,
-  pageNumberStyle,
-  swiperContainer,
-} from "./PopularPots.style";
+import { buttonContainer, swiperContainer } from "./PopularPots.style";
 import { PotCard } from "@components/index";
 import useGetPots from "apis/hooks/pots/useGetPots";
 import { useState } from "react";
-import { LeftIcon } from "@assets/svgs";
 import Skeleton from "react-loading-skeleton";
 import { cardStyle } from "@components/cards/PotCard/PotCard.style";
 import "react-loading-skeleton/dist/skeleton.css";
+import PaginationButton from "@components/commons/Pagination/Pagination";
 
 const PopularPots = () => {
   const [currentPage, setCurrentPage] = useState(1);
@@ -28,16 +22,12 @@ const PopularPots = () => {
   const totalPages = Math.ceil(pots.length / 3);
   const currentPots = pots.slice((currentPage - 1) * 3, currentPage * 3);
 
-  const handleNextPage = () => {
-    if (currentPage < totalPages) {
-      setCurrentPage(currentPage + 1);
-    }
+  const handlePrev = () => {
+    if (currentPage > 1) setCurrentPage((prev) => prev - 1);
   };
 
-  const handlePrevPage = () => {
-    if (currentPage > 1) {
-      setCurrentPage(currentPage - 1);
-    }
+  const handleNext = () => {
+    if (currentPage < totalPages) setCurrentPage((prev) => prev + 1);
   };
 
   return (
@@ -80,25 +70,12 @@ const PopularPots = () => {
                   </SwiperSlide>
                 ))}
             <div css={buttonContainer}>
-              <button
-                type="button"
-                css={buttonStyle}
-                onClick={handlePrevPage}
-                disabled={currentPage === 1}
-              >
-                <LeftIcon />
-              </button>
-              <p css={pageNumberStyle}>
-                {currentPage} / {totalPages}
-              </p>
-              <button
-                type="button"
-                css={buttonStyle}
-                onClick={handleNextPage}
-                disabled={currentPage === totalPages}
-              >
-                <LeftIcon css={iconStyle} />
-              </button>
+              <PaginationButton
+                currentPage={currentPage}
+                totalPages={totalPages}
+                onPrev={handlePrev}
+                onNext={handleNext}
+              />
             </div>
           </Swiper>
         </>
