--- src/pages/MyPotDetail/pages/MyPotStatus/components/Pagination/Pagination.tsx
@@ -1,21 +1,0 @@
-import { paginationContainer, paginationTextStyle } from "../../MyPotStatus.style";
-import { ArrowButton } from "@components/index";
-
-interface PaginationProps {
-  currentPage: number;
-  totalPages: number;
-  onPrev: () => void;
-  onNext: () => void;
-}
-
-const Pagination: React.FC<PaginationProps> = ({ currentPage, totalPages, onPrev, onNext }) => (
-  <div css={paginationContainer}>
-    <ArrowButton direction="left" onClick={onPrev} />
-    <span css={paginationTextStyle}>
-      {currentPage} / {totalPages}
-    </span>
-    <ArrowButton direction="right" onClick={onNext} />
-  </div>
-);
-
-export default Pagination;
