+++ src/pages/MyPotDetail/pages/MyPotStatus/components/index.ts
@@ -2,7 +2,7 @@ import MyTodoModal from "./MyTodoModal/MyTodoModal";
 import MyPotTodoCard from "./MyPotTodoCard/MyPotTodoCard";
 import MyPotTodoList from "./MyPotTodoList/MyPotTodoList";
 import StatusBar from "./StatusBar/StatusBar";
-import Pagination from "./Pagination/Pagination";
+import Pagination from "../../../../../components/commons/Pagination/Pagination";
 import TaskCardList from "./TaskCardList/TaskCardList";
 import TodoStatusSection from "./TodoStatusSection/TodoStatusSection";
